+++ src/ddzy/__tests__/utilityAlgorithm.test.ts
@@ -3,6 +3,7 @@ import {
   ListNode,
 } from "../utility/algorithm/double-linked-circular-list/list-node";
 
+
 describe('utilityAlgorithm tests', () => {
   // ? Binary Search Tree
   describe('BST', () => {
@@ -1044,6 +1045,162 @@ describe('utilityAlgorithm tests', () => {
         }, '');
 
         expect(result).toBe(expected);
+      });
+    });
+
+    // ? _filter
+    describe('ES6Achieve._filter tests', () => {
+      const _filter = utilityAlgorithm.ES6Achieve._filter;
+
+      test('_filter should return a new empty array when receive an empty array', () => {
+        const received: number[] = [];
+
+        const result = _filter<number>(received, (v) => {
+          return !!v;
+        });
+
+        expect(result.length).toBe(0);
+        received.push(2);
+        received.push(3);
+        expect(received.length).toBe(2);
+        expect(result.length).toBe(0);
+      });
+
+      test('_filter should return the eligible array when receive an array composed of number', () => {
+        const received: number[] = [1, 2, 3, 4, 5, 6, 7, 8];
+        const expected: number[] = [2, 4, 6, 8];
+
+        const result = _filter<number>(received, (v) => {
+          return v % 2 === 0;
+        });
+
+        for (const [i, v] of result.entries()) {
+          expect(v).toBe(expected[i]);
+        }
+      });
+
+      test('_filter should return the eligible array when receive an array composed of plain object', () => {
+        interface User {
+          name: string,
+          age: number,
+        };
+        const received: User[] = [
+          { name: 'duan', age: 20 },
+          { name: 'zhao', age: 30 },
+          { name: 'yang', age: 40 },
+        ];
+        const expected: User[] = [
+          { name: 'zhao', age: 30 },
+          { name: 'yang', age: 40 },
+        ];
+
+        const result = _filter<User>(received, (v) => {
+          return v.age >= 30;
+        });
+
+        for (const [i, v] of result.entries()) {
+          expect(v.name).toBe(expected[i].name);
+          expect(v.age).toBe(expected[i].age);
+        }
+      });
+
+      test('_filter should return the aligible array and print the truthy context when receive an array composed of number', () => {
+        interface Obj {
+          secret: string,
+          say: (secret: string) => void,
+        };
+
+        const obj: Obj = {
+          secret: '980808',
+          say() {
+            console.log(this.secret);
+          },
+        };
+        const received = [-1, 1, -2, 2, -3, 3];
+        const expected = {
+          context: obj,
+          arr: [1, 2, 3],
+        };
+
+        const result = _filter<number, Obj>(received, function (v) {
+          expect(this).toBe(expected.context);
+
+          return v > 0;
+        }, obj);
+
+        for (const [i, v] of result.entries()) {
+          expect(v).toBe(expected.arr[i]);
+        }
+      });
+    });
+
+    // ? _every
+    describe('ES6Achieve._every tests', () => {
+      const _every = utilityAlgorithm.ES6Achieve._every;
+
+      test('_every should always return `true` when receive an empty array', () => {
+        const received: number[] = [];
+        const expected = true;
+
+        const result = _every<number>(received, (v) => {
+          return v > 0;
+        });
+
+        expect(result).toBe(expected);
+      });
+
+      test('_every should return `true` when receive an array composed of `number` that accord with condition', () => {
+        const received: number[] = [1, 2, 3, 4, 5];
+        const expected = true;
+
+        const result = _every<number, null>(received, (v) => {
+          return v < 10;
+        });
+
+        expect(result).toBe(expected);
+      });
+
+      test('_every should return `false` when receive an array composed of `object` that not accord with condition', () => {
+        interface IPair {
+          name: string;
+          age: number;
+        };
+
+        const received: IPair[] = [
+          { name: 'duan', age: 21 },
+          { name: 'zhao', age: 31 },
+          { name: 'yang', age: 41 },
+        ];
+        const expected = false;
+
+        const result = _every<IPair, null>(received, (v) => {
+          return v.age < 0;
+        });
+
+        expect(result).toBe(expected);
+      });
+
+      test('_every should called by custom `this` context', () => {
+        const context = {
+          name: 'ddzy',
+          printName() {
+            return this.name;
+          },
+        };
+
+        const received: number[] = [1, 2, 3, 4, 5];
+        const expected = {
+          context,
+          result: true,
+        };
+
+        const result = _every<number, typeof context>(received, function (v) {
+          expect(this).toBe(expected.context);
+
+          return v < 6;
+        }, context);
+
+        expect(result).toBe(expected.result);
       });
     });
   });